minor update to db parsing

+++ ExecutorBase/Game/ClientCard.cs
@@ -304,12 +304,12 @@ namespace WindBot.Game
         public bool HasSetcode(int setcode)
         {
             if (Data == null) return false;
-            long setcodes = Data.Setcode;
+            ulong setcodes = Data.Setcode;
             int settype = setcode & 0xfff;
             int setsubtype = setcode & 0xf000;
             while (setcodes > 0)
             {
-                long check_setcode = setcodes & 0xffff;
+                long check_setcode = (long)(setcodes & 0xffff);
                 setcodes >>= 16;
                 if ((check_setcode & 0xfff) == settype && (check_setcode & 0xf000 & setsubtype) == setsubtype) return true;
             }
